# frontend/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable error code",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EventLevel": {
        "enum": [
          "DEBUG",
          "INFO",
          "WARNING",
          "ERROR"
        ],
        "title": "EventLevel",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunCreated": {
        "properties": {
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          }
        },
        "required": [
          "run_id",
          "status",
          "task_id"
        ],
        "title": "RunCreated",
        "type": "object"
      },
      "RunDetail": {
        "properties": {
          "applications_prepared": {
            "title": "Applications Prepared",
            "type": "integer"
          },
          "applications_submitted": {
            "title": "Applications Submitted",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cv_generated": {
            "title": "Cv Generated",
            "type": "integer"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error_count": {
            "title": "Error Count",
            "type": "integer"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/RunError"
            },
            "title": "Errors",
            "type": "array"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RunEventOut"
            },
            "title": "Events",
            "type": "array"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "jobs_discovered": {
            "title": "Jobs Discovered",
            "type": "integer"
          },
          "jobs_processed": {
            "title": "Jobs Processed",
            "type": "integer"
          },
          "jobs_qualified": {
            "title": "Jobs Qualified",
            "type": "integer"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "run_type": {
            "$ref": "#/components/schemas/RunType"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "summary": {
            "additionalProperties": true,
            "title": "Summary",
            "type": "object"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "trigger": {
            "$ref": "#/components/schemas/RunTrigger"
          }
        },
        "required": [
          "id",
          "run_type",
          "trigger",
          "status",
          "created_at",
          "started_at",
          "finished_at",
          "duration_seconds",
          "jobs_discovered",
          "jobs_processed",
          "jobs_qualified",
          "cv_generated",
          "applications_prepared",
          "applications_submitted",
          "error_count",
          "task_id",
          "errors",
          "parameters",
          "summary",
          "events"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunError": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "type",
          "message",
          "at"
        ],
        "title": "RunError",
        "type": "object"
      },
      "RunEventOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "level": {
            "$ref": "#/components/schemas/EventLevel"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "id",
          "sequence",
          "created_at",
          "level",
          "stage",
          "message",
          "data"
        ],
        "title": "RunEventOut",
        "type": "object"
      },
      "RunPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "RunPage",
        "type": "object"
      },
      "RunStatus": {
        "enum": [
          "PENDING",
          "RUNNING",
          "SUCCEEDED",
          "PARTIAL_SUCCESS",
          "FAILED",
          "CANCELLED"
        ],
        "title": "RunStatus",
        "type": "string"
      },
      "RunSummary": {
        "properties": {
          "applications_prepared": {
            "title": "Applications Prepared",
            "type": "integer"
          },
          "applications_submitted": {
            "title": "Applications Submitted",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "cv_generated": {
            "title": "Cv Generated",
            "type": "integer"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error_count": {
            "title": "Error Count",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "jobs_discovered": {
            "title": "Jobs Discovered",
            "type": "integer"
          },
          "jobs_processed": {
            "title": "Jobs Processed",
            "type": "integer"
          },
          "jobs_qualified": {
            "title": "Jobs Qualified",
            "type": "integer"
          },
          "run_type": {
            "$ref": "#/components/schemas/RunType"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "trigger": {
            "$ref": "#/components/schemas/RunTrigger"
          }
        },
        "required": [
          "id",
          "run_type",
          "trigger",
          "status",
          "created_at",
          "started_at",
          "finished_at",
          "duration_seconds",
          "jobs_discovered",
          "jobs_processed",
          "jobs_qualified",
          "cv_generated",
          "applications_prepared",
          "applications_submitted",
          "error_count",
          "task_id"
        ],
        "title": "RunSummary",
        "type": "object"
      },
      "RunTrigger": {
        "enum": [
          "MANUAL",
          "SCHEDULED",
          "API",
          "SYSTEM"
        ],
        "title": "RunTrigger",
        "type": "string"
      },
      "RunType": {
        "enum": [
          "DIAGNOSTIC",
          "DAILY_PIPELINE",
          "DISCOVERY",
          "ANALYSIS",
          "CV_GENERATION",
          "APPLICATION_PREPARATION",
          "SUBMISSION"
        ],
        "title": "RunType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "API token (API_AUTH_TOKEN)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\nSelf-hosted AI job application agent.\n\n* **Discovery \u2192 analysis \u2192 CV generation \u2192 application preparation** are automated.\n* **Submission** always requires explicit approval by default (`AUTO_SUBMIT=false`).\n* All endpoints except `/health/*` require `Authorization: Bearer <API_AUTH_TOKEN>`.\n",
    "title": "AI Job Application Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/runs": {
      "get": {
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "run_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List automation runs (newest first)",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/diagnostic": {
      "post": {
        "operationId": "start_diagnostic_api_v1_runs_diagnostic_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The task queue is unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start a system self-test run on a worker",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/discovery": {
      "post": {
        "operationId": "start_discovery_api_v1_runs_discovery_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The task queue is unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start a job discovery run on a worker",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run detail with its event timeline",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}